requirements.txt
Flask
Flask-SQLAlchemy
pytz
fpdf2
qrcode
Pillow
gunicorn